#35- BindingAdapter 파일 삭제 및 통합
